requirements.txt
google-genai
requests
SpeechRecognition
pyttsx3
pyaudio